dcp-815 split up metadata unittests
